actualizacion de apis

+++ src/main/java/org/pi/Services/CategoriaService.java
@@ -4,6 +4,8 @@ import org.pi.Repositories.CategoriaRepository;
 
 import java.sql.SQLException;
 import java.util.List;
+import java.util.NoSuchElementException;
+
 //falta la logica de las reglas del negocio
 public class CategoriaService {
     private final CategoriaRepository categoriaRepository;
@@ -16,18 +18,55 @@ public class CategoriaService {
         return categoriaRepository.findAllCategoria();
     }
     public Categoria findCategoria(int id) throws SQLException{
-        return categoriaRepository.findCategoria(id);
+        if (id <= 0){
+            throw new IllegalArgumentException("La id debe ser mayor a cero");
+        }
+        Categoria categoria = categoriaRepository.findCategoria(id);
+        if (categoria == null){
+            throw new NoSuchElementException("La categoria no existe");
+        }
+        return categoria;
     }
 
-    public void saveCategoria(Categoria categoria) throws SQLException{
-        categoriaRepository.saveCategoria(categoria);
+    public int saveCategoria(Categoria categoria) throws SQLException{
+        String nombre = categoria.getNombreCategoria();
+
+        List<Categoria> categorias = findAllCategoria();
+        boolean existe = categorias.stream().anyMatch(c -> c.getNombreCategoria().equalsIgnoreCase(nombre));
+
+        if (existe){
+            throw new IllegalArgumentException("Ya existe una categoria con ese nombre");
+        }
+        return  categoriaRepository.saveCategoria(categoria);
     }
 
     public void deleteCategoria(int id) throws SQLException{
+        if (id <= 0){
+            throw new IllegalArgumentException("La id debe ser mayor a cero");
+        }
+        Categoria categoria = categoriaRepository.findCategoria(id);
+        if (categoria == null){
+            throw new NoSuchElementException("No se puede eliminar la categoria. No se encontro la categoria");
+        }
         categoriaRepository.deleteCategoria(id);
     }
 
     public void updateCategoria( Categoria categoria) throws SQLException{
+        if (categoria.getIdCategoria() <= 0){
+            throw new IllegalArgumentException("La id debe ser mayor a cero");
+        }
+        Categoria existe = categoriaRepository.findCategoria(categoria.getIdCategoria());
+        if (existe == null){
+            throw new NoSuchElementException("No se puede actualizar la categoria. No se encontro la categoria");
+        }
+        String nombre = categoria.getNombreCategoria();
+        List<Categoria> categorias = findAllCategoria();
+        boolean nombreExiste = categorias.stream().anyMatch(c -> c.getNombreCategoria().equalsIgnoreCase(nombre)
+                && c.getIdCategoria() != categoria.getIdCategoria());
+
+        if (nombreExiste){
+            throw new IllegalArgumentException("Otra categoria ya tiene ese nombre");
+        }
         categoriaRepository.updateCategoria(categoria);
     }
 }
